An alternative to handling Runnables
This PR is meant to highlight how I would reorganize the classes such that the Runnable doesn't instantiate a thread itself but instead defers to the caller. This also makes `stop() a synchronous call (though we could revert this portion or make it configurable). It also gets rid of the `start()` methods entirely.
+ Made `WriteConnThread` a `Runnable` that has a synchronous `stop()` method
+ Made `ServerWorkerThread` a `Runnable` that has a synchrnous `stop()` method. Furthermore, it launches a WriteConnThread only when a `ServerWorkerThread#run()` method is executed
+ Made `ServerLoadBalancer` use a fixed thread pool to execute the workers